+++ src/defineReact.js
@@ -47,7 +47,7 @@ function defineReact(react, noLongerUsed, extend)
 	};
 	
 	// equivalent of `extends React.Component` or other class if provided via `extend` param
-	ext(RefluxComponent, _extend);
+	Reflux.utils.inherits(RefluxComponent, _extend);
 	
 	proto = RefluxComponent.prototype;
 	
@@ -316,6 +316,18 @@ function defineReact(react, noLongerUsed, extend)
 		configurable: true
 	});
 	
+	// allows a shortcut for accessing MyStore.singleton.state as MyStore.state (since common usage makes a singleton)
+	Object.defineProperty(RefluxStore, "state", {
+		get: function () {
+			if (!this.singleton) {
+				throw new Error('Reflux.Store.state is inaccessible before the store has been initialized.');
+			}
+			return this.singleton.state;
+		},
+		enumerable: true,
+		configurable: true
+	});
+	
 	/* NOTE:
 	If a Reflux.Store definition is given a static id property and used
 	properly within a Reflux.Component or with Reflux.initStore then
@@ -434,19 +446,6 @@ function filterByStoreKeys(storeKeys, obj)
 		}
 	}
 	return doUpdate ? updateObj : false;
-}
-
-// used as a well tested way to mimic ES6 class `extends` in ES5 code
-function ext(d, b) {
-    for (var p in b) {
-		if (b.hasOwnProperty(p)) {
-			d[p] = b[p];
-		}
-	}
-    function __() {
-		this.constructor = d;
-	}
-    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
 }
 
 // this is utilized by some of the global state functionality in order to get a clone that will